% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{adaptedData}
\alias{adaptedData}
\title{Convenience function for returning adapted data}
\usage{
adaptedData(x, train = TRUE)
}
\arguments{
\item{x}{Object of class \code{fairadapt}, a result of an adaptation
procedure.}

\item{train}{A logical indicating whether train data should be returned.
Defaults to \code{TRUE}. If \code{FALSE}, test data is returned.}
}
\description{
Convenience function for returning adapted data
}
