% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fairadapt.R
\docType{package}
\name{fairadapt-package}
\alias{fairadapt-package}
\title{fairadapt: Fair Data Adaptation with Quantile Preservation}
\description{
An implementation of the fair data adaptation with quantile preservation described in Plecko & Meinshausen (JMLR 2020, 21(242), 1-44). The adaptation procedure uses the specified causal graph to pre-process the given training and testing data in such a way to remove the bias caused by the protected attribute. The procedure uses tree ensembles for quantile regression. Instructions for using the methods are further elaborated in the corresponding JSS manuscript, see \doi{10.18637/jss.v110.i04}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/dplecko/fairadapt}
  \item Report bugs at \url{https://github.com/dplecko/fairadapt/issues}
}

}
\author{
\strong{Maintainer}: Drago Plecko \email{www.plecko@gmail.com}

Authors:
\itemize{
  \item Nicolas Bennett \email{nicolas.bennett@stat.math.ethz.ch}
}

}
\keyword{internal}
