% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fair_Adaptation.R
\name{FairAdapt}
\alias{FairAdapt}
\title{FairAdapt}
\usage{
FairAdapt(formula, train.data, test.data, adj.mat, protect.A,
  res.vars = NULL)
}
\arguments{
\item{formula}{Object of class \code{formula} describing the response and the covariates.}

\item{train.data, test.data}{Training data & testing data, both of class \code{data.frame}.}

\item{adj.mat}{Matrix of class \code{matrix} encoding the relationships in the causal graph. \code{M[i,j] == 1} implies the existence of an edge from node i to node j. Must include all the}

\item{protect.A}{A value of class \code{character} describing the binary protected attribute. Must be one of the entries of \code{colnames(adj.mat)}.}

\item{res.vars}{A vector of class \code{character} listing all the resolving variables, which should not be changed by the adaption procedure. Default value is \code{NULL}, corresponding to no resolving variables. Resolving variables should be a subset of the descendants of the protected attribute.}
}
\value{
A \code{list} of length two. The two elements of the list are of class \code{data.frame} and contain the adapted training and testing data respectively.
}
\description{
Implementation of fair adaptation procedures (Plecko & Meinshausen 2019).
Uses only plain \code{R}.
}
\details{
The procedure takes the training and testing data as an input, together with the causal graph given by an adjacency matrix and the list of resolving variables, which should be kept fixed during the adaptation procedure. The procedure then calculates a fair representation of the data, after which any classification method can be used. There are, however, several valid training options yielding fair predictions, and the best of them can be chosen with cross-validation. For more details we refer the user to the original paper.
Most of the running time is due to the quantile regression step using the ranger package. There should soon be a C++ code patch to enhance performance of quantile regression in ranger, directly improving the performance of fairadapt.
}
\examples{
\donttest{
library(fairadapt)

# UCI Adult example
L <- FairAdapt(income ~ ., train.data = adult.train,
               test.data = adult.test, protect.A = "sex",
               adj.mat = adjacency.matrix)
adjusted.train.data <- L[[1]]
adjusted.test.data <- L[[2]]
}

}
\references{
Plecko, D. & Meinshausen, N. (2019). Fair Adaptation Procedures \cr
}
\author{
Drago Plecko
}
