% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{plot.fairadapt}
\alias{plot.fairadapt}
\title{Plotting data before and after adaptation}
\usage{
\method{plot}{fairadapt}(x, when = "after", ...)
}
\arguments{
\item{x}{An object of class \code{"fairadapt"}.}

\item{when}{A \code{character(1L)} indicating whether data before or after adaptation
is visualized. Default value is \code{"after"}, but can also take values \code{"before"}
and \code{"both"} (in which case both visualizations are provided).}

\item{...}{In this case ignored.}
}
\value{
A base R plot for visualizing the distribution of the outcome
before/after the adaptation procedure.
}
\description{
Plotting data before and after adaptation
}
