% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{quantFit}
\alias{quantFit}
\title{Quality of quantile fit statistics.}
\usage{
quantFit(x, ...)
}
\arguments{
\item{x}{Object of class \code{fairadapt}, a result of an adaptation
procedure.}

\item{...}{Ignored in this case.}
}
\value{
A \code{numeric} vector, containing the average empirical loss for
of the 25%, 50% and 75% quantile loss functions, for each variable.
}
\description{
Quality of quantile fit statistics.
}
\examples{
uni.adj.mat <- array(0, dim = c(4, 4))
colnames(uni.adj.mat) <- rownames(uni.adj.mat) <-
  c("gender", "edu", "test", "score")

uni.adj.mat["gender", c("edu", "test")] <-
  uni.adj.mat["edu", c("test", "score")] <-
  uni.adj.mat["test", "score"] <- 1L

FA <- fairadapt(score ~ .,
  train.data = uni_admission[1:100, ],
  test.data = uni_admission[101:150, ],
  adj.mat = uni.adj.mat, prot.attr = "gender", eval.qfit = 3L)

quantFit(FA)
}
