% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quants.R
\name{rangerQuants}
\alias{rangerQuants}
\alias{linearQuants}
\alias{mcqrnnQuants}
\title{Quantile Engine Constructor for the Quantile Learning step.}
\usage{
rangerQuants(data, A.root, ind, min.node.size = 20, ...)

linearQuants(
  data,
  A.root,
  ind,
  tau = c(0.001, seq(0.005, 0.995, by = 0.01), 0.999),
  ...
)

mcqrnnQuants(
  data,
  A.root,
  ind,
  tau = seq(0.005, 0.995, by = 0.01),
  iter.max = 500,
  ...
)
}
\arguments{
\item{data}{A \code{data.frame} with data to be used for quantile
regression.}

\item{A.root}{A \code{logical(1L)} indicating whether the protected
attribute \code{A} is a root node of the causal graph. Used for splitting the
quantile regression.}

\item{ind}{A \code{logical} vector of length \code{nrow(data)}, indicating 
which
samples have the baseline value of the protected attribute.}

\item{min.node.size, ...}{Forwarded to \link[ranger]{ranger}.}

\item{tau, ...}{Forwarded to \code{\link[quantreg]{rq}}.}

\item{tau, iter.max, ...}{Forwarded to \link[qrnn]{mcqrnn.fit}.}
}
\value{
A \code{ranger} or a \code{rangersplit} S3` object, depending on the 
value of the \code{A.root} argument, for \code{rangerQuants()}.

A \code{rqs} or a \code{quantregsplit} S3 object, depending on the 
value of the \code{A.root} argument, for \code{linearQuants()}.

An \code{mcqrnn} S3 object for \code{mcqrnnQuants()}.
}
\description{
There are several functions that can be used for the quantile learning step
in the \code{fairadapt} package.
}
\details{
The \code{rangerQuants()} function uses random forests 
(\code{ranger} package) for quantile regression.

The \code{linearQuants()} function uses linear quantile regression 
(\code{quantreg} package) for the Quantile Learning step.

The \code{mcqrnnQuants()} function uses  monotone quantile 
regression neural networks (`mcqrnn` package) in the Quantile Learning step.
}
