% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{summary.fairadapt}
\alias{summary.fairadapt}
\alias{print.summary.fairadapt}
\title{Summarizing fairadapt fit}
\usage{
\method{summary}{fairadapt}(object, ...)

\method{print}{summary.fairadapt}(x, digits = max(3L, getOption("digits") - 3L), ...)
}
\arguments{
\item{object}{An object of class \code{"fairadapt"}.}

\item{...}{In this case ignored.}

\item{x}{Object of class \code{"summary.fairadapt"}.}

\item{digits}{Number of digits appearing in the output.}
}
\value{
Summary of the object formula, protected attribute, attribute levels,
resolving variables, number of training and test samples, adapted variables,
TV measure before adaptation, TV measure after adaptation, and the quantile
method that was used.
}
\description{
\code{summary} method for class \code{"fairadapt"}.
}
