% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boot.R
\name{summary.fairadaptBoot}
\alias{summary.fairadaptBoot}
\alias{print.summary.fairadaptBoot}
\title{Summarizing fairadaptBoot fit}
\usage{
\method{summary}{fairadaptBoot}(object, ...)

\method{print}{summary.fairadaptBoot}(x, ...)
}
\arguments{
\item{object}{An object of class \code{"fairadaptBoot"}.}

\item{...}{In this case ignored.}

\item{x}{Object of class \code{"summary.fairadaptBoot"}.}
}
\value{
Summary of the bootstrap wrapper call, protected attribute,
attribute levels,
resolving variables, number of training and test samples, adapted variables,
number of bootstrap repetitions, indicator if the quantileFit objects were
saved, randomness mode, and the quantile method that was used.
}
\description{
\code{summary} method for class \code{"fairadaptBoot"}.
}
