% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{uni_admission}
\alias{uni_admission}
\title{Univeristy admission data of 1,000 students.}
\format{
A data frame with 1,000 rows and 4 variables:
\describe{
  \item{gender}{the gender of the student}
  \item{edu}{educational achievement, for instance GPA}
  \item{test}{performance on a university admission test}
  \item{score}{overall final score measuring the quality of a candidate}
}
}
\usage{
uni_admission
}
\description{
A simulated dataset containing the evaluation of students' abilities.
}
\keyword{datasets}
