% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{visualizeGraph}
\alias{visualizeGraph}
\title{Visualize Graphical Causal Model}
\usage{
visualizeGraph(x, ...)
}
\arguments{
\item{x}{Object of class \code{fairadapt}, a result of an adaptation
procedure.}

\item{...}{Additional arguments passed to the graph plotting function.}
}
\description{
Visualize Graphical Causal Model
}
