% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fairadapt.R
\docType{package}
\name{fairadapt-package}
\alias{fairadapt-package}
\alias{_PACKAGE}
\title{fairadapt: Fair Data Adaptation with Quantile Preservation}
\description{
An implementation of the fair data adaptation with quantile preservation described in Plecko & Meinshausen (2019) <arXiv:1911.06685>. The adaptation procedure uses the specified causal graph to pre-process the given training and testing data in such a way to remove the bias caused by the protected attribute. The procedure uses tree ensembles for quantile regression.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/dplecko/fairadapt}
  \item Report bugs at \url{https://github.com/dplecko/fairadapt/issues}
}

}
\author{
\strong{Maintainer}: Drago Plecko \email{drago.plecko@stat.math.ethz.ch}

Authors:
\itemize{
  \item Nicolas Bennett \email{nicolas.bennett@stat.math.ethz.ch}
}

}
\keyword{internal}
